Exclude the "daily" database for now
A lambda layer must have an _uncompressed_ size of less than 256MiB, and with both the "main"
and "daily" databases included we don't get there. For testing having the "main" database is
already good (and better than having to download that as well!), in the future we can work on
further improving this.