Use the CodeBuild build number as _patch_ version component
The previous logic had two flaws: `-bXXX` wouldn't be a "build" but rather is treated
as a "prerelease", and as such would be considered "before" the base version and never match
the intended "~1.0.0" spec.
At the same time: IFF it had been a build version (M.M.P+B), then the semver spec actually
says to ignore it always in any comparison, so that wouldn't help us either.

This commit changes the version so that the developer specifies a major.minor version in the
`BASE_VERSION` environment variable, and the build number is actually used to produce the
patch version component. In this way we clearly communicate "this is important", and the build
number is a reasonably good patch version in that it goes up. As it never goes down we may
see "holes" in the version ranges when bumping major/minor, but we can solve that at that point
by either simply ignoring that (1.0.250 would be followed by 1.1.251 for example), or by adding
a corrective offset to bring things in line.